Consolidate auth flows into single login page
Merge login and signup into a unified auth page at /login with mode switching via query params. Extracted form logic into reusable components (LoginForm, SignUpForm), enhanced auth tab styling with transitions and hover effects, and removed separate signup/entrar routes in favor of redirects.
