+++ apps/web/app/_components/LandingNav.tsx
@@ -1,7 +1,6 @@
 'use client';
 import { useState } from 'react';
 import Link from 'next/link';
-import { ArrowRight } from 'lucide-react';
 import { VelarBrand } from '../../components/VelarBrand';
 
 const NAV_ITEMS = [
@@ -45,7 +44,7 @@ export function LandingNav() {
         </nav>
 
         <a
-          href="/ir-login"
+          href="/login"
           className="group inline-flex h-11 cursor-pointer items-center gap-2 rounded-full bg-primary px-5 text-[14px] font-semibold text-white no-underline transition hover:bg-primary-container hover:shadow-lg hover:shadow-primary/25"
         >
           Acceder a la plataforma
