+++ apps/web/app/globals.css
@@ -291,11 +291,22 @@ body {
   min-height: 38px;
   align-items: center;
   justify-content: center;
+  cursor: pointer;
   border-bottom: 1px solid rgba(255, 255, 255, 0.16);
   color: rgba(226, 235, 252, 0.62);
   font-size: 1rem;
   font-weight: 500;
   text-decoration: none;
+  transition:
+    color 0.18s ease,
+    border-color 0.18s ease,
+    transform 0.18s ease;
+}
+
+.velar-auth-tab:hover {
+  border-bottom-color: rgba(91, 155, 255, 0.5);
+  color: #e8f1ff;
+  transform: translateY(-1px);
 }
 
 .velar-auth-tab[aria-current="page"] {
