+++ apps/web/app/signup/page.tsx
@@ -1,11 +1,5 @@
-import type { Metadata } from 'next';
-import SignUpPageClient from './SignUpPageClient';
-
-export const metadata: Metadata = {
-  title: 'Crear cuenta | VELAR',
-  description: 'Registra una cuenta en VELAR para participar en la plataforma de trazabilidad.',
-};
+import { redirect } from 'next/navigation';
 
 export default function SignUpPage() {
-  return <SignUpPageClient />;
+  redirect('/login?mode=signup');
 }
