+++ apps/web/middleware.ts
@@ -1,6 +1,5 @@
 import { NextResponse, type NextRequest } from 'next/server';
 import { createServerClient } from '@supabase/ssr';
-import { getSafeRedirectTarget } from './lib/auth/routing';
 
 export async function middleware(request: NextRequest) {
   let supabaseResponse = NextResponse.next({ request });
@@ -32,10 +31,6 @@ export async function middleware(request: NextRequest) {
     const loginUrl = new URL('/login', request.url);
     loginUrl.searchParams.set('next', pathname + request.nextUrl.search);
     return NextResponse.redirect(loginUrl);
-  }
-  if (user && pathname === '/login') {
-    const nextTarget = getSafeRedirectTarget(request.nextUrl.searchParams.get('next'));
-    return NextResponse.redirect(new URL(nextTarget ?? '/', request.url));
   }
 
   return supabaseResponse;
